+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_internalnetwork.py
@@ -39,9 +39,9 @@ def step_create(test, checks=None):
     if checks is None:
         checks = []
     test.cmd('az networkfabric internalnetwork create --resource-group {rg} --l3-isolation-domain-name {l3domain} --resource-name {name} --vlan-id {vlan_id}'
-             ' --mtu {mtu} --extension {extension} --is-monitoring-enabled {isMonitoringEnabled} --connected-ipv4-subnets {connectedIpv4Subnets}'
+             ' --mtu {mtu} --connected-ipv4-subnets {connectedIpv4Subnets}'
              ' --static-route-configuration {staticRouteConf} --bgp-configuration {bgpConf}'
-             ' --import-route-policy {importRoutePolicy} --export-route-policy {exportRoutePolicy}', checks=checks)
+             ' --import-route-policy {importRoutePolicy}', checks=checks)
 
 
 def step_update(test, checks=None):