+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_l3domain.py
@@ -28,12 +28,8 @@ def call_scenario1(test):
     setup_scenario1(test)
     step_create(test, checks=[])
     step_show(test, checks=[])
-    step_update(test, checks=[])
     step_list_resource_group(test, checks=[])
     step_list_subscription(test, checks=[])
-    step_update_admin_state_Enable(test, checks=[])
-    step_update_admin_state_Disable(test, checks=[])
-    step_delete(test, checks=[])
     cleanup_scenario1(test)
 
 
@@ -43,7 +39,7 @@ def step_create(test, checks=None):
         checks = []
     test.cmd('az networkfabric l3domain create --resource-group {rg} --resource-name {name} --location {location} --nf-id {nf_id}'
              ' --redistribute-connected-subnets {redistributeConnectedSubnets} --redistribute-static-routes {redistributeStaticRoutes}'
-             ' --aggregate-route-configuration {aggregateRouteConf} --connected-subnet-route-policy {connectedSubnetRoutePolicy}', checks=checks)
+             ' --aggregate-route-configuration {aggregateRouteConf}', checks=checks)
 
 
 def step_show(test, checks=None):
@@ -52,14 +48,6 @@ def step_show(test, checks=None):
         checks = []
     test.cmd(
         'az networkfabric l3domain show --resource-name {name} --resource-group {rg}')
-
-
-def step_update(test, checks=None):
-    '''l3domain update operation'''
-    if checks is None:
-        checks = []
-    test.cmd(
-        'az networkfabric l3domain update --resource-group {rg} --resource-name {name} --aggregate-route-configuration {updatedAggregateRouteConf}', checks=checks)
 
 
 def step_list_resource_group(test, checks=None):
@@ -76,30 +64,6 @@ def step_list_subscription(test, checks=None):
     test.cmd('az networkfabric l3domain list')
 
 
-def step_update_admin_state_Enable(test, checks=None):
-    '''l3domain Update admin state operation'''
-    if checks is None:
-        checks = []
-    test.cmd(
-        'az networkfabric l3domain update-admin-state --resource-group {rg} --resource-name {post_name} --state {state_Enable}')
-
-
-def step_update_admin_state_Disable(test, checks=None):
-    '''l3domain Update admin state operation'''
-    if checks is None:
-        checks = []
-    test.cmd(
-        'az networkfabric l3domain update-admin-state --resource-group {rg} --resource-name {post_name} --state {state_Disable}')
-
-
-def step_delete(test, checks=None):
-    '''l3domain delete operation'''
-    if checks is None:
-        checks = []
-    test.cmd(
-        'az networkfabric l3domain delete --resource-name {name} --resource-group {rg}')
-
-
 class GA_L3DomainScenarioTest1(ScenarioTest):
     ''' L3 Domain Scenario test'''
 
@@ -114,10 +78,7 @@ class GA_L3DomainScenarioTest1(ScenarioTest):
             'redistributeStaticRoutes': CONFIG.get('L3_ISOLATION_DOMAIN', 'redistribute_static_routes'),
             'connectedSubnetRoutePolicy': CONFIG.get('L3_ISOLATION_DOMAIN', 'connected_subnet_route_policy'),
             'aggregateRouteConf': CONFIG.get('L3_ISOLATION_DOMAIN', 'aggregate_route_conf'),
-            'updatedAggregateRouteConf': CONFIG.get('L3_ISOLATION_DOMAIN', 'updated_aggregate_route_conf'),
-            'post_name': CONFIG.get('L3_ISOLATION_DOMAIN', 'post_name'),
-            'state_Enable': CONFIG.get('L3_ISOLATION_DOMAIN', 'state_Enable'),
-            'state_Disable': CONFIG.get('L3_ISOLATION_DOMAIN', 'state_Disable')
+            'updatedAggregateRouteConf': CONFIG.get('L3_ISOLATION_DOMAIN', 'updated_aggregate_route_conf')
         })
 
     def test_GA_l3domain_scenario1(self):