+++ src/quantum/azext_quantum/operations/job.py
@@ -382,9 +382,9 @@ def _submit_directly_to_service(cmd, resource_group_name, workspace_name, locati
     if storage is None:
         from .workspace import get as ws_get
         ws = ws_get(cmd)
-        if ws.storage_account is None:
+        if ws.properties.storage_account is None:
             raise RequiredArgumentMissingError("No storage account specified or linked with workspace.")
-        storage = ws.storage_account.split('/')[-1]
+        storage = ws.properties.storage_account.split('/')[-1]
     job_id = str(uuid.uuid4())
     container_name = "quantum-job-" + job_id
     connection_string_dict = show_storage_account_connection_string(cmd, resource_group_name, storage)
@@ -470,7 +470,7 @@ def _submit_qsharp(cmd, program_args, resource_group_name, workspace_name, locat
     """
     Submit a Q# project to run on Azure Quantum.
     """
-    _show_warning('The direct submission of Q# programs will soon be deprecated. Instead, you can submit QIR bitcode or LLVM. Modern QDK can be used to generate LLVM from Q#.')
+    _show_warning('The direct submission of Q# project folders will soon be fully deprecated. Instead, you can submit QIR bitcode or human-readable LLVM code. Modern QDK can be used to generate human-readable LLVM code from Q#.')
     # We first build and then call run.
     # Can't call run directly because it fails to understand the
     # `ExecutionTarget` property when passed in the command line